feat: remove char width limitation
This commit removes the limitation that `char` width must be less than
or equal to 1 and instead now enforces that `char` must be a non-empty
string (width > 0).

Previously empty string is allowed as value for `char`. While this
does not causes any issue, it (IMO) does not add any value either.
Since the purpose of this plugin is to display a char in place of
colorcolumn, it really doesn't make sense to render a non-displayable
character, hence the enforcement for non-empty string.

Removing the maximum limit of char width also open new possibility,
user can now choose a multi-width char (or even a full string) as
character to display.

Signed-off-by: Jakka Prihatna <jprihatna@gmail.com>